Fix duplicate metrics in multiprocess mode
Prometheus client multiprocessing mode disables registries isolation,
so all metrics are exported. Multiprocess collector is configured
on one registry. If any metrics are explicitly connected to that
registry they end up in the metrics output twice - once for the
aggregate multiprocess value and once with the values from the
individual worker process. To avoid this we don't connect any metrics
to the registry and use the default one instead.