Use request route expression instead of path
[Prometheus docs][cardinality-docs] recommend to keep the metric
cardinality low and avoid using unique identifiers in the labels.

Depending on the application, using request path as a metric label
could dramatically increase the amount of data stored.

To avoid this, we're using request route pattern instead of the actual
path as the metric label (eg `/documents/<id>` instead of `/documents/123`).

This still allows to differentiate between different request types but
since the number of endpoints in an app is generally limited it avoids
creating thousands (or millions) of additional time series.

[cardinality-docs]: https://prometheus.io/docs/practices/instrumentation/#do-not-overuse-labels