Make teardown function not throw an uncaught AttributeError
Problem:
We can not guarantee that `g` has the `_gds_metrics_start_time`
attribute. This is because it is set in the `before_request` function
but we can't guarantee that that function will be run. This is because
if in another `before_request` function that throws an uncaught
exception before ours then no other `before_request` functions are run
but the `teardown_request` function still is, throwing an
`AttributeError`.

Solution:
Our code now handles the case were `_gds_metrics_start_time` is not set
and still registers the metric for the request count but does not try
for the request duration (as it does not know the duration).

It is also encouraged that when users of this library call `init_app`,
they do so before calling `init_app` for other libraries. This will mean
that our `before_request` function is called before any others.